+++ src/main/java/racing_car/step4/controller/RacingGame.java
@@ -2,6 +2,7 @@ package racing_car.step4.controller;
 
 import racing_car.step4.domain.Cars;
 import racing_car.step4.domain.CarsFactory;
+import racing_car.step4.domain.RandomMoveStrategy;
 import racing_car.step4.input.InputView;
 import racing_car.step4.output.ResultView;
 
@@ -15,7 +16,7 @@ public class RacingGame {
         ResultView.resultMessagePrint();
         
         while (numberOfTry-- > 0) {
-            cars = new Cars(cars.moveTry());
+            cars = new Cars(cars.tryMove(new RandomMoveStrategy()));
             ResultView.processPrint(cars);
         }
     }
