+++ src/test/java/racing_car/step4/domain/PositionTest.java
@@ -2,7 +2,6 @@ package racing_car.step4.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racing_car.step3.domain.Position;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -11,6 +10,6 @@ class PositionTest {
     @Test
     @DisplayName("포지션 증가 확인")
     void increase() {
-        assertThat(new racing_car.step3.domain.Position(1)).isEqualTo(new Position(0).increase());
+        assertThat(new Position(1)).isEqualTo(new Position(0).increase());
     }
 }
